refactor(app): extract log toggling
